renamed enums -> definitions
