+++ test/binary-parser.test.js
@@ -8,7 +8,7 @@ const { encodeAccountID } = require('ripple-address-codec')
 const { binary } = require('../dist/coretypes')
 const { Amount, Hash160 } = coreTypes
 const { makeParser, readJSON } = binary
-const { Field, TransactionType } = require('./../dist/enums')
+const { Field, TransactionType } = require('./../dist/definitions')
 const { parseHexOnly, hexOnly, loadFixture } = require('./utils')
 const { bytesToHex } = require('../dist/utils/bytes-utils')
 const fixtures = loadFixture('data-driven-tests.json')
